resolving comments, moved propose form to another path
